Add HhResumeService and tests
