+++ src/main/java/org/example/autobot/service/HhResumeService.java
@@ -4,6 +4,7 @@ import org.example.autobot.model.HhProfile;
 import org.example.autobot.repository.HhProfileRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
 import org.springframework.stereotype.Service;
@@ -13,11 +14,16 @@ import reactor.core.publisher.Mono;
 @Service
 public class HhResumeService {
     private static final Logger log = LoggerFactory.getLogger(HhResumeService.class);
-    private final WebClient client = WebClient.create("https://hh.ru");
+    private final WebClient client;
     private final HhProfileRepository repo;
 
-    public HhResumeService(HhProfileRepository repo) {
+    public HhResumeService(
+            HhProfileRepository repo,
+            WebClient.Builder builder,
+            @Value("${hh.base-url:https://hh.ru}") String baseUrl
+    ) {
         this.repo = repo;
+        this.client = builder.baseUrl(baseUrl).build();
     }
 
     /**
